requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-jalali==6.0.1
django-modeltranslation==0.18.11
django-rest-auth==0.9.5
django-utils-six==2.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
jdatetime==4.1.1
kavenegar==1.1.2
Pillow==10.1.0
PyJWT==2.8.0
python-decouple==3.8
python-gettext==5.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
